add reqwest and seperate test folder for integ tests
- add reqwest
- add openssl lib and tell pkg_config where to find it
- create tests dir for integration tests
